requirements.txt
attrs
pyserial